+++ lib/linkhut_web/components/core_components.ex
@@ -99,7 +99,7 @@ defmodule LinkhutWeb.CoreComponents do
     default: nil,
     doc: "custom error translator function, defaults to translate_error/1"
 
-  attr :rest, :global, include: ~w(autocomplete cols disabled form max maxlength min minlength
+  attr :rest, :global, include: ~w(accept autocomplete cols disabled form max maxlength min minlength
                                    pattern placeholder readonly required rows size step)
   slot :inner_block