+++ lib/linkhut_web/router.ex
@@ -158,6 +158,7 @@ defmodule LinkhutWeb.Router do
     get "/archive/:link_id", SnapshotController, :show
     get "/archive/:link_id/all", SnapshotController, :index
     post "/archive/:link_id/recrawl", SnapshotController, :recrawl
+    post "/archive/:link_id/upload", SnapshotController, :upload
     get "/archive/:link_id/snapshot/:id/delete", SnapshotController, :remove_snapshot
     delete "/archive/:link_id/snapshot/:id", SnapshotController, :delete
     get "/archive/:link_id/:format/full", SnapshotController, :full