+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to linkhut will be documented in this file.
 - Add option to strip tracking parameters (UTM, fbclid, gclid, etc.) from new bookmark URLs
 - Add S3-compatible storage backend for snapshots
 - Add support for deleting individual snapshots
+- Add support for uploading snapshots
 
 ## [0.1.2] - 2026-03-12